dev-games/godot: add 4.0_beta1, keyworded
Was already planning to keyword once it's out of alpha with the
first beta given realistically new projects should be started
on 4 even if beta (migration from 3 is non-trivial).

Reminder that it's slotted and 3 is very much still supported, so
godot:3 can be added to world file for working on or running older
projects.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>