dev-util/mingw64-toolchain: fix lto again
Was wondering why AR was looking for the lto plugin in a non-existing
bfd-plugins directory before, turned out it was because gcc resolved
its location differently when set as a symlink.

In commit c4262506ff492b96cddccb15e1fe1842d8d5a626, reverted to using
upstream's intended hardlink but didn't notice this change in behavior
and it broke lto again.

Bug: https://bugs.gentoo.org/854516
Fixes: c4262506ff492b96cddccb15e1fe1842d8d5a626
Signed-off-by: Ionen Wolkens <ionen@gentoo.org>