app-emulation/vkd3d: drop broken live ebuild
There is little use for this package anymore given recent Wine
versions bundle it to ease handling mingw-built PE files and do
not use the system's ELF libraries.

In-tree wine-7.0 can still optionally use it so it's not going
anywhere yet, but fixing and keeping the live ebuild for stable
releases doesn't seem useful.

Note that if interested in the latest features from Valve, these
are included in vkd3d-proton-9999 instead and can be used by any
Wine versions (similarly to dxvk by adding dlls to the Wine prefix).

Bug: https://bugs.gentoo.org/685918
Signed-off-by: Ionen Wolkens <ionen@gentoo.org>