media-libs/libpng: fix NEON on ARM64
As of 1.6.38, =check is banned on arm64 because NEON is always present
there. We don't need =check otherwise, so just enable conditionally
based on whether cpu_flags_arm_neon is on.

(configure also says =check is deprecated/unsupported.)

Note that while cpu_flags_arm_neon is masked on arm64 in general,
it's unmasked for packages which we know work, like libpng (as
it's been fine before 1.6.38).

Closes: https://bugs.gentoo.org/870298
Thanks-to: Roy Bamford <neddyseagoon@gentoo.org>
Signed-off-by: Sam James <sam@gentoo.org>