net-libs/nodejs: force libgcc as support lib
Upstream behavior is correct wrt libatomic (unconditionally adding it)
due to their use of atomic_is_lock_free which is yet to be added to
compiler-rt (stuck in review hell at https://reviews.llvm.org/D85044).

This necessitates forcing libgcc as support lib.  Linking to libatomic
from gcc is dynamic, so sys-devel/gcc has to go in RDEPEND.  Clang can
be forced to use libgcc with --rtlib=libgcc, even with
USE=default-compiler-rt (which is the original cuase of the bug in
question).

    This applies this pull request to 14.x and 16.x.

Closes: https://github.com/gentoo/gentoo/pull/27271
Signed-off-by: William Hubbs <williamh@gentoo.org>