sys-devel/autoconf: backport K&R decls fix to 2.13
Backport the K&R decls fix to 2.13 to avoid configure tests
failing (often "silently", i.e. doesn't fail the build of
the package overall, just leads to wrong results) with
newer compilers like the upcoming Clang 16.

A consequence of this whole fuss is that we're going to
have to eautoreconf in a bunch of older packages, but
as Ionen pointed out on IRC, this means we get some
other stuff for free like Python 3.10 fixes, so it's
not all bad.

Undecided how to handle any packages with a generated
configure from autoconfs not in tree. We may just
patch the configure manually. Will see.

See linked bug 870412 and the Discourse thread therein for
details if unfamiliar.

As I remarked on IRC, it's unfortunate to have to backport anything
which will then infect generated configure, as it's somewhat
a repeat of the --runstatedir situation, and it leads to confusion/hassle
when sending patches upstream where a generated/bootstrapped configure
is required.

But this is different - it's a (serious) bug being fixed
rather than a gratuitous (although) understandable given we weren't
sure if we'd ever see another autoconf release backport of a new
feature.

Bug: https://bugs.gentoo.org/806376
Bug: https://bugs.gentoo.org/870412
Thanks-to: Arsen Arsenovic <arsen@aarsen.me>
Signed-off-by: Sam James <sam@gentoo.org>