x11-drivers/nvidia-drivers: drop unnecessary patches
Ended up using -Werror=strict-prototypes to test things, but as far
as I'm aware this shouldn't be "needed" with clang16 either unless
something passes that -Werror (so kernel module still need these
fixes). Not that it'd hurt to fix but don't want to maintain patches
on this end. Will revisit if needed.

nvidia is aware of modules issues regarding this, so clang15 patches
should be removable later

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>