media-gfx/tuxpaint: fix LIBDIR, fix doc on 0.9.28
The LIBDIR variable defined while calling emake
collides with the LIBDIR variable defined in the tuxpaint
Makefile, resulting in magic tools not getting linked
properly.
Renaming LIBDIR to GENTOO_LIBDIR resolves this issue.

Furthermore, a patch for the DOC_PREFIX variable in the
Makefile was missing, resulting in installation of files
into unexpected paths.

Bug: https://bugs.gentoo.org/873286
Bug: https://bugs.gentoo.org/870232
Signed-off-by: Philipp Rösner <rndxelement@protonmail.com>
Closes: https://github.com/gentoo/gentoo/pull/27518
Signed-off-by: Joonas Niilola <juippis@gentoo.org>