net-libs/libpcap: add test
There's not really a test suite upstream (see mentioned link).  There's
a handful of "test programs" (which we at least test building and
linking with this change), but this one (findalldevstest) is the only
one that is actually run (under valgrind) in upstream CI.  On the
upside, it should be rather reproducible since only the loopback
interface will ever be exposed inside the portage network sandbox.

See: https://github.com/the-tcpdump-group/libpcap/issues/1012
Bug: https://bugs.gentoo.org/705802
Signed-off-by: Matoro Mahri <matoro@users.noreply.github.com>