sys-apps/flashrom: do not depend on virtual/libusb:0
There was only a single programmer which still depended on that version
of libusb, and it builds fine against virtual/libusb:1. In fact, even
before this change upstream build scripts happily ignored libusb-0 and
only linked against libusb-1 even with pickit2-spi support enabled!

Straight-to-stable revbump because as mentioned above, the dependency in
question was not actually used.

Closes: https://bugs.gentoo.org/873769
Signed-off-by: Marek Szuba <marecki@gentoo.org>