games-fps/yamagi-quake2: adjust opengl/gles deps
Mesa is the vendor which could be something else (aka nvidia), and
not much sense in checking directly. On the other hand, I needed to
enable gles2 on libsdl2 (flag handles gles3 too) or it'd complain
when I tried to use it, or at least with libsdl2[-opengl].

virtual/opengl is also mostly deprecated, so switch to libglvnd (this
is not wired to use libOpenGL over libGL so still need [X]).

Also add gles as a REQUIRED_USE option, I could USE="client gles -*"
with libsdl2[-opengl,gles2] and run this with +set vid_renderer gles3.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>