dev-python/PyQt6: add 6.4.0
qt-6.4.0 is only in the ::qt overlay the moment, but this works
just fine with 6.3.2 that I can see. Adding but may be better to
wait to match 6.x as a precaution in the future, notably confuses
1 qutebrowser test but it has no real consequence.

Eitherway, given it's masked not too woried about this and want
to get some testing out of these things to know effects of mismatch.

PyQt6-6.4.0 adds support for QtTextToSpeech and QtPdf but neither
can be enabled at the moment.
- pdf: from qtwebengine:6 but is currently disabled in its ebuild
- speech: not available in <qt-6.4.0 and qtspeech:6 is not packaged
  in the qt overlay either yet, skips the need to bump QT_PV to
  6.4:6 here for now

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>