java-pkg-simple.eclass: improve MANIFEST.MF manipulation
For adding Main-class and/or Automatic-module-name to MANIFEST.MF it
seems better to use "jar" functionality at the end of compilation.
This does the necessary checks and issues warnings in case of duplicate
entries.
Erroneously used existence checks ( -v ) were causing wrong entries in
MANIFEST.MF files in case of JAVA_MAIN_CLASS="" which is often used in
multi-jar ebuilds.

Closes: https://bugs.gentoo.org/874924
Signed-off-by: Volkmar W. Pogatzki <gentoo@pogatzki.net>