sys-libs/libsmbios: fix incorrect KEYWORDS
Mark as -* until upstream is properly fixed to not require <sys/io.h>
unconditionally. The ~arm64 and ~riscv keywords were added by mistake;
they both fail to build due to absence of said header.

Based on glibc availability of <sys/io.h>, alpha is the only other
architecture that can build the package in any case, but SMBIOS is not
relevant on alpha anyway so ~alpha is not tested or included.

Closes: https://bugs.gentoo.org/875830
Closes: https://github.com/gentoo/gentoo/pull/27672
Signed-off-by: WANG Xuerui <xen0n@gentoo.org>
Signed-off-by: Sam James <sam@gentoo.org>