dev-vcs/git-filter-repo: Move to distutils-r1, bump to -r1
The new b4 depends on git-filter-repo as a module, not just as a script.
This converts git-filter-repo over to distutils-r1, via two hacks:

* The package provides a setup.py, but it's not in a standard location.
* The github packages don't provide a PKG-INFO so distutils can't find
  the version, this just makes a minimal one inline.

With those I can get git-filter-repo installed as both a script and a
module.  I've bumped the version of these to -r1 so users get a rebuild
with the new files.

Signed-off-by: Palmer Dabbelt <palmer@rivosinc.com>