app-crypt/coolkey: fix build w/ upcoming clang16
Header that declare this is not installed, and seems like these
headers would conflict in all sorts of ways here either way.

An alternative fix could be to drop --enable-pk11install which is
marked legacy (off by default), but don't know this package and
no idea if anyone need this.

Closes: https://bugs.gentoo.org/871585
Signed-off-by: Ionen Wolkens <ionen@gentoo.org>