dev-games/godot: add 4.0_beta3
This changes things around for target and tools flags (latter
is removed and replaced by target=editor), and dev_build replaces
target=debug. Adjusted so it hopefully works the same as before.

Please report if anything off, these alpha/beta change a lot of
things around and it's not always obvious to keep up with scons
combined with the long build times.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>