dev-db/mysql-connector-c++: fix loading libmysqlclient.so
mysql-workbench would fail with mysql-connector-c++ 8.x because it
would try to load the obsolete libmysqlclient_r library. This version
no longer exists in newer versions of mysql-connector-c. It turns out
that mysql-connector-c++ has hard-coded references to this library
but since we don't install compatibility symlinks this hardcoded
reference will fail.

Closes: https://bugs.gentoo.org/831664
Signed-off-by: Hans de Graaff <graaff@gentoo.org>