app-portage/g-octave:
This patch will modify g-octave so that the ebulds it generates will be compliant with EAPI 7.
The ebuilds will now use PATCHES instead of adding epatch lines to the src_prepare stage.
OCTLIBDIR is now passed to the linker to allow mkoctfile to build against the correct location of the octave libraries.
Modified DB_DIR to match the folder name inside the current db archive.
No longer die'ing in pkg_config if there are no old files to remove.
Patching the info.json to use the new sys-libs/ncurses-compat:5 for libncurses5-dev dependencies instead of sys-libs/ncurses which is now at version 6.

Signed-off-by: Kevin Martin <kevinm@gatech.edu>