net-misc/icaclient: RDEPEND on sys-libs/llvm-libunwind
Some binaries need libunwind.so.1 and installing sys-libs/llvm-libunwind
used to be an issue because it caused a conflict with net-fs/samba.
Plus it was unknown if the llvm version was the correct one, or if
icaclient in fact needed an old sys-libs/libunwind.

Now it is clear that sys-libs/llvm-libunwind is correct.
https://docs.citrix.com/en-us/citrix-workspace-app-for-linux/configure-xenapp.html#adding-the-libunwind-12-library-dependency-for-llvm-12

Closes: https://bugs.gentoo.org/855374
Signed-off-by: Henning Schild <henning@hennsch.de>
Closes: https://github.com/gentoo/gentoo/pull/27867
Signed-off-by: Joonas Niilola <juippis@gentoo.org>