dev-python/PyQt6: match IUSE defaults with qtbase further
Feel this is the generally sane way to handle defaults
as don't get far without base modules.

This adds opengl and printsupport on top of the rest, both
already used by qutebrowser (albeit still misses qml which
needs a non-base module).

Exception for testlib as it's afaik useless outside tests.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>