dev-python/libvirt-python: Drop nonexistent useflag mask
This dates back to the day when app-emulation/libvirt used to
have 'python' use flag because it used to bundle python bindings.
As it was transitioning into a separate repository/separate
ebuild the new ebuild was made to conflict with
app-emulation/libvirt[python]. But this is all ancient history
and app-emulation/libvirt doesn't have the python use flag
anymore. Therefore, drop it from RDEPEND check.

Signed-off-by: Michal Privoznik <michal.privoznik@gmail.com>