sys-libs/libxcrypt: drop multibuild patch for xcrypt.h, add verify-sig
The multibuild patch causes a fair inconvenience on bumps because
it required a patch which touches the build system. libxcrypt
is a dependency of Perl which is a dependency of autotools, so
we can't eautoreconf within the ebuild. That means every bump
has to be accompanied with a manual 'make dist' tarball
from a Gentoo developer.

I don't think we need to install xcrypt.h -- we want
to keep prebuilt binaries working if they need the old
libxcrypt symbols (which I suspect barely anything used anyway,
it's not the same as glibc libcrypt compatibility), but
supporting building new binaries against that API is not something
we're interested in.

Signed-off-by: Sam James <sam@gentoo.org>