sys-devel/autoconf: further Clang 16 patches for autoconf 2.13
Reported by matsl when building XEmacs 21.4 (21.5 - which is WIP - works
with newer autoconf).

A bunch of the built-in/provided tests from autoconf 2.13 suffered
from -Wimplicit-int or -Wimplicit-function-declaration. We also
disable the {get,set}pgrp tests (see the patch for details, but
the gist is they're unfixable without some substantial effort
on POSIX systems nowadays and there's no need anyway -- they're
gone in newer autoconf and we never expect them to fail.)

Bug: https://bugs.gentoo.org/874366
Reported-by: Mats Lidell <matsl@gentoo.org>
Signed-off-by: Sam James <sam@gentoo.org>