app-misc/remind: Change SRC_URI to Debian's GitLab
Upstream has changed how they store tarballs in a way that breaks
ebuilds: they only maintain the latest release on the download path,
while all older versions are moved to an OLD subdirectory. This means
that whenever a new release is cut it the latest ebuild will fail to
download if not using a Gentoo distfiles mirror.

As per the application's web page the Debian GitLab is an official
mirror. Switching over to it means predictable download URLs for each
release. It also allows for using tar.bz2 instead of tar.gz, which saves
disk space and bandwidth.

Signed-off-by: Zoltan Puskas <zoltan@sinustrom.info>
Signed-off-by: Sam James <sam@gentoo.org>