media-libs/libplacebo: add 5.229.1
With partly revamped ebuild, and:
* ensure use right python executble wrt bug #731728
* disable demos wrt bug #851927, there's little use in enabling
  as only one of them gets installed by default and it's a very
  barebone test video player

USE=glslang is now a no-op if shaderc is set, these provide
the same functionality and shaderc is the default pick upstream.
Previously, glslang would force disable shaderc even if set.
Alternatively could use REQUIRED_USE to prevent setting both
but feel this is mostly an annoyance.

Closes: https://bugs.gentoo.org/731728
Closes: https://bugs.gentoo.org/851927
Co-authored-by: Nick Sarnie <sarnex@gentoo.org>
Signed-off-by: Ionen Wolkens <ionen@gentoo.org>