media-video/mpv: add 0.34.1_p20221105
mpv releases are rare and it can be difficult to backport fixes
(e.g. old libplacebo is broken with newer glslang, but new
libplacebo breaks current mpv, not that can't disable glslang
or vulkan on mpv if need be).

So will release snapshots if felt necessary, please report if feel
a newer snapshot would be useful.

One quirk is that a whole year ago it switched to API 2.0 and not
all revdeps are ready for this (mostly a few trivial deprecations).
Personally see this as a good time to get this sorted out.

Note this also include meson migration, which may still have
oversights. MacOS / rpi builds are also untested, feel free to
make any changes to improve these two if you use/know them.

rough list of non-style changes from -9999 (non-exhaustive):
* update DESCRIPTION, upstream been dissociating itself with the
  original MPlayer
* drop pkg_postinst versions messages (0.18-0.25 is rather ancient)
* drop docdir patch as it installed examples .conf to /usr/share/mpv,
  instead just move these to docdir/examples
* drop IUSE=cplugins and enable by default, upstream default and
  not seeing a motivation to be optional
* drop IUSE=gbm, rather redundant with IUSE=drm and is default on mesa
* limit USE=opengl to controlling gl-x11/gl-cocoa backends rather than
  the general `gl` option, avoids e.g. `egl? ( opengl )` req use
* make IUSE=opengl non-default given gl-x11 is deprecated, albeit
  still let it be enabled by desktop profiles
* add IUSE=+libplacebo split off from IUSE=vulkan and depend on newer
  version, if recent enough this enables `-vo gpu-next` which is not
  limited to vulkan (default given libplacebo main author is same as
  mpv's and this is the obvious focus)
* invert IUSE=debug logic, was passing N(no)DEBUG with debug
* swap mesa with libglvnd for IUSE=egl, low level driver packages
  handling is not for us (and mesa is meaningless with nvidia with
  the exception of getting libgbm)
* ^ likewise, drop nvidia-drivers (ffmpeg/obs/etc.. also don't check
  drivers for nvenc)
* only depend on docutils + build man/html with USE=cli, and drop
  IUSE=doc (mpv.html has same dependency as the man page)
* fix python-exec[-native-symlinks] when USE=-tools, calls 'python3'
  at build time too
* depend on vulkan-loader[X?,wayland?] wrt bug #753434 and #777774
* various deps and tentative REQUIRED_USE adjustments

Closes: https://bugs.gentoo.org/753434
Closes: https://bugs.gentoo.org/777774
Signed-off-by: Ionen Wolkens <ionen@gentoo.org>