media-gfx/hydrus: remove direct dependency on media-video/mpv
mpv itself does not provide python bindings and using PYTHON_USEDEP
on it is meaningless and will cause issues now that mpv is migrating
to python-single-r1 (only used for 1 /usr/bin script).

As far as I can tell this package does not use mpv directly and
always goes through python-mpv, so let it handle pulling libmpv.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>