media-video/vidcutter: use mpv slot, backport api 2.0 fix
libmpv.so.1 -> libmpv.so.2

opengl_cb been deprecated for ~3 years and is now gone, this
backports a rough update to pympv (thus the large patch).
*looks* fine but I don't use this to really know.

Unlike other packages, this also breaks compatibility with
stable libmpv at runtime (due to a version check). And upstream
also seem to have went back & forth with this (broken again
on current master). Thus the conditional patching for now.
Since this changes nothing for stable, went straight-to.

At same time also ${P} -> ${PF} for docdir as revisions are
installing docs in the wrong place.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>