media-fonts/x11fonts-jmk: Install files with 0644 permissions
font.eclass has some racy code in pkg_postinst() that changes
permissions of already-installed files. I want to remove that to avoid
the race. This is the only package that installs fonts with permissions
other than 0644, so override that in src_install().

The claim in font.eclass is that fontconfig segfaults if fonts are
unreadable, but that claim dates to 2007 (bug #187774). Additionally,
0444 is readable, but who knows. Let's just keep things working how they
have been since 2007.

Bug: https://bugs.gentoo.org/187774
Signed-off-by: Matt Turner <mattst88@gentoo.org>