sci-physics/bullet: rename altivec vector keywords
Altivec vectors can be defined with either the "vector" keyword or the
"__vector" type. In general "__vector" should be prefered for
include files, as otherwise it might conflicts with other type define
in the source code (define a vector class in C++ is quite common).
This causes bullet to fail to build on powerpc if the
code is compiled with -maltivec, or by default on ppc64el which always
has altivec enabled.

See also: https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=760310
Closes: https://bugs.gentoo.org/852101
Signed-off-by: Niccolò Belli <niccolo.belli@linuxsystems.it>
Closes: https://github.com/gentoo/gentoo/pull/28193
Signed-off-by: Andrew Ammerlaan <andrewammerlaan@gentoo.org>