sys-devel/clang-common: drop -Werror=deprecated-non-prototype for 15.x, promote to 16.x
Clang 16.x will make -Wimplicit-int, -Wimplicit-function-declaration, and
-Wincompatible-function-pointer-type error by default.

Clang 15.x+ also warns on -Wdeprecated-non-prototype but it's not becoming
fatal yet. It'll be an error when the defaults in compilers change to C23.

There's a LOT of breakage with -Werror=deprecated-non-prototype, and we do
care about it, but the first batch of errors mentioned above are more
important for the time being. We need to be able to triage and prioritise
the bugs.

So:
* Clang 15 USE=stricter: only -Wimplict-int -Wimplicit-function-declaration
  and -Wincompatible-function-pointer-type are errors.

* Clang 16 USE=stricter: upstream defaults (which include ^ as we're
  doing for clang 15) + -Wdeprecated-non-prototype as an error, because presumably
  if you set USE=stricter on *16*, you want something harsher than already
  will ship upstream.

This more accurately lets developers in Gentoo set USE=stricter on clang-common
for 15.x and get Clang 16.x behaviour rather than scaring the life out of them.

Thanks to Ionen for talking this out with me.

Bug: https://bugs.gentoo.org/876985
See: 930fe2268cbedda61e37fada65e57352d25d8761
Signed-off-by: Sam James <sam@gentoo.org>