media-sound/easyeffects: (R)DEPEND on sci-libs/gsl
Needed alongside media-libs/speex (which by the way is technically
an *optional* dependency but given Easy Effects also have
a mandatory dependency on media-libs/speexdsp, adding IUSE=speex
would IMHO be confusing) since 7.0.0 but I forgot to paste that one
in.

With thanks to Xian Wang for having spotted this before me.

Closes: https://bugs.gentoo.org/880745
Signed-off-by: Marek Szuba <marecki@gentoo.org>