media-libs/sdl-ttf: add 2.0.11_p20220525
Upstream aren't making releases anymore (since a long time ago!) for the 1.2.x
branch but are kindly doing backports, so let's make a snapshot.

The vulnerable (CVE-2022-27470) code doesn't seem to be in 1.2.x - and
given upstream are quite good about backporting, the absence of any related
commits seems to support that.

Bug: https://bugs.gentoo.org/843434
Signed-off-by: Sam James <sam@gentoo.org>