app-shells/bash: backport another extglob fix
It's still not sufficient but with this in on top of _p9, it's very
clear what the remaining issues are then, rather than me at least
trying to remember exactly what's broken still as the saga continues.

It is still not enough to unmask as there's still one extglob
issue remaining (and this came up pretty early as far as testing
goes). I don't see us being able to expose Bash 5.2 to users
for a while yet.

Bug: https://bugs.gentoo.org/873931
Signed-off-by: Sam James <sam@gentoo.org>