profiles/arch/loong: remove the media-sound/jack2[ieee1394] package.use.mask
Now media-libs/libffado is confirmed to build on loong, there should
be no further arch-specific reasons for ieee1394 to not work on loong
despite me having no hardware to test on. Unmask for now; someone can
always yell in the various IM channels/rooms/groups to have the mask
restored in case anything breaks for them.

Signed-off-by: WANG Xuerui <xen0n@gentoo.org>