x11-terms/kitty: adjust live in advance for verify-sig
Mostly so don't forget to handle this. If upstream is going to vendor
Go deps, then this should work given go-module_src_unpack will be a
no-op, otherwise will have to switch to verify-sig_verify_detached
so can add our vendor tarball to SRC_URI and also not unpack twice.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>