x11-terms/kitty: adjust live further for go-module.eclass
This is not fully a Go package, there's just a helper tool written in
Go. But go-module sets some broad assumptions that everything is Go
even if we were to use GO_OPTIONAL (it's tempting to simply not use
the eclass, only need to set a few variables + BDEPEND if we're not
going to use EGO_SUM).

Also quote EPYTHON while here, technically should never have spaces per
definition but let's not rely on this too much for eclasses.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>