dev-lang/R: add 4.2.2
My upstream patches to the R_BROWSER and R_PDFVIEWER search order didn't
land in this release, but we're already running eautoreconf, so we might
as well include them. This lets us drop the dependency on xdg-utils, and
remove the stanza (conditional on USE=X) that sets the default browser
and viewer to xdg-open.

We are also able to drop a patch to the test suite that was merged
upstream.

Closes: https://bugs.gentoo.org/458176
Closes: https://bugs.gentoo.org/879863
Signed-off-by: Michael Orlitzky <mjo@gentoo.org>