profiles: unmask Bash 5.2 and Readline 8.2
There's still a few issues - their presence is subideal, but there's
no specific regressions wrt Gentoo usage now, and they're things
to take to upstream (and lobby them over) if it's a problem.

That is, there's some changes in behaviour the Bash community is
understandably not happy about, but there's nothing concrete which
upstream would agree is a bug as things stand.

Bug: https://bugs.gentoo.org/881379
Closes: https://bugs.gentoo.org/873931
Signed-off-by: Sam James <sam@gentoo.org>