kde-frameworks/kwindowsystem: Add IUSE wayland to pull in runtime plugin
All other plugins from kde-plasma/kwayland-integration went into respective
Frameworks packages and only the kwindowsystem plugin remains, which makes
this possible to depend on. It is necessary because kwindowsystem consumers
will need the wayland runtime plugin even on non-Plasma systems.

This is not exactly pretty but the same exception as dev-qt/qtgui[wayland].

Bug: https://bugs.gentoo.org/790275
Signed-off-by: Andreas Sturmlechner <asturm@gentoo.org>