media-sound/xmms2: switch to waf-utils.eclass + adjustments
Was unused when picked up this package and left it like that,
but might as well use it to be consistent with the recent eclass
changes (albeit calling waf-utils_src_compile in src_test feels
a bit dodgy to avoid calling it directly).

Doing straight to stable, this is scarcely used and not too worried.

Also adjust deps a bit and filter-lto.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>