dev-libs/gmp: pass --enable-fat for x86 (includes more asm)
Vanilla (_not_ from an ebuild) gmp will probe CHOST and then
deduce its *own* tuple based on the currently used CPU, abusing
its own custom config.guess (which our econf overwrites). This
value is used to decide which accelerated asm is compiled in (MPN_PATH).

In ebuilds, we override and refresh config.guess and friends because
it's a common portability problem to have stale versions which don't
even recognise e.g. riscv as an arch. But gmp's non-standard use
of these files poses a problem for us.

We can't simply resort to just not refreshing/changing the files
because we need to be able to build binary packages on a system
without them being super specific to the machine they were built
on with no way of controlling it (we may want to look at a USE
flag for this going forward though).

For now, we can use --enable-fat which produces a generic libgmp
which can load various asm at runtime based on the CPU in use - success!

Unfortunately, --enable-fat only does anything on x86 for the time
being, so !x86 is still in the same situation as before this commit,
but this is progress at least.

Thanks goes to Nikita Zlobin <nick87720z@gmail.com> for noticing
the problem originally (and being persistent with investigating it).

Temporarily unkeyworded until I can test this a bit more.

Bug: https://bugs.gentoo.org/883201
Signed-off-by: Sam James <sam@gentoo.org>