net-analyzer/nagios-plugins: new revision to trigger a rebuild.
Anyone whose "ping" has moved will want to rebuild this after the latest
fix. Doing so right now won't help everybody (e.g. if nagios-plugins is
upgraded before iputils), but it's an easy rebuild and will help *some*
people, so I think it's worth it. Thanks to Sam for the suggestion.

Bug: https://bugs.gentoo.org/883765
Signed-off-by: Michael Orlitzky <mjo@gentoo.org>