net-analyzer/monitoring-plugins: new revisions to fix ping path.
We pass explicit ping/ping6 commands to the monitoring-plugins
./configure script to prevent it from running those commands during the
build. However, instead of grabbing their paths from $PATH, we have
(until now) hard-coded them to /bin/ping. This has now bitten us with
net-misc/iputils-20221126.

This commit uses $(command -v ...) to get the right absolute
paths. Thanks to Daniel Pouzzner for reporting the problem and
suggesting the near-identical fix to nagios-plugins. I've made the
changes and revisions to both the ~arch and stable ebuilds to avoid as
much breakage as possible, but we will still need to coordinate the
stabilization of monitoring-plugins-2.3.2-r1 with iputils-20221126.

Closes: https://bugs.gentoo.org/883729
Signed-off-by: Michael Orlitzky <mjo@gentoo.org>