net-misc/icaclient: bump to 22.12.0.12
Here we do not just bump the version but also change the way we deal
with the shared config file between this ebuild and its plugin
media-plugins/hdx-realtime-media-engine.
Also we move the rdep to sys-libs/llvm-libunwind to the plugin. And if
the user wants hdx support on icaclient they have to set a USE flag. The
plugin will not work if icaclient is not built with support for it and
the plugin itself contains the tool to enable itself so we turn the dep
chain around, the plugin is even a bdep of the application.

Closes: https://bugs.gentoo.org/883775
Closes: https://bugs.gentoo.org/881637
Signed-off-by: Henning Schild <henning@hennsch.de>
Closes: https://github.com/gentoo/gentoo/pull/28493
Signed-off-by: Sam James <sam@gentoo.org>