app-emulation/wine-proton: destabilize 7.0.5 back to ~x86
Stabilized wine-vanilla-7.0.1 was fine, but upon closer look
wine-proton on the other hand is kinda broken at runtime when
doing any X11 usage. Keeping keyword given believe may be because
of my x86 test setup, but at least going to avoid stable.

Not that I really want x86 either way, mostly was to be in sync
with other wine versions.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>