app-alternatives/yacc: install env.d file for YACC
Without this, autotools will end up probing for bison even when
it's not specifically needed.

If packages really do need bison, it's not expected that
ebuild maintainers will fix the underlying problem, but
the ebuilds should then set YACC=bison and BDEPEND on it.

If upstream is active, they should consider reporting the problem
to them though.

Signed-off-by: Sam James <sam@gentoo.org>