app-alternatives/lex: install env.d file for LEX
Without this, autotools will end up probing for flex even when
it's not specifically needed.

If packages really do need flex, it's not expected that
ebuild maintainers will fix the underlying problem, but
the ebuilds should then set LEX=flex and BDEPEND on it.

If upstream is active, they should consider reporting the problem
to them though.

Signed-off-by: Sam James <sam@gentoo.org>