net-dns/djbdns: new revision increasing tinydns's softlimit.
The tinydns launch script imposes a "softlimit" on the amount of memory
it can use, and sometime recently (in djbdns terms, anyway), the old
default of 300000 became insufficient. We raise it 4500000. This will
only help new installations; users with existing ones will need to
modify their local launch scripts because the script itself is not
controlled by the package manager.

Thanks to Vadim Efimov for pointing out that this problem is widespread
and in need of fixing.

Closes: https://bugs.gentoo.org/883159
Signed-off-by: Michael Orlitzky <mjo@gentoo.org>