dev-libs/tree-sitter-cpp: restrict tests
Requires sample files from tree-sitter-c to be available, but these
files are not installed with dev-libs/tree-sitter-c, and are instead
expected to be pulled down via making the tree-sitter-c javascript
wrapper package a dependency in package.json.  However since we invoke
"tree-sitter test" directly rather than via "npm test" the source
package for tree-sitter-c will not be available.

Bug: https://bugs.gentoo.org/844223
Signed-off-by: Matoro Mahri <matoro@users.noreply.github.com>