app-alternatives/yacc: don't set YACC for Bison for now
Leaving this for now to be safe, as it's closer to pre-alternatives
status quo to leave it unset and let autoconf probe for Bison by itself
as it prefers it anyway, and might be a CPP-like situation wrt
calling bison or bison -y if YACC is set.

Thanks to Ionen again.

See: 8d975a71210fb429ace5a8fd5e2d8dcbd52b4339
Signed-off-by: Sam James <sam@gentoo.org>