app-dicts/*: Apply patsub_replacement defences
Per bug #881383, string replacing forms of parameter expansion must take care
to quote - or appropriately escape - any nested parameter expansions, assuming
that their values are intended to be taken literally (as is almost invariably
the case). This has long been the case, but the introduction of the new
patsub_replacement option in bash >=5.2 has brought the issue to the fore.

In the case of the app-dicts package category, the improper quoting is not yet
causing any issues. Still, it is better to write the code properly to begin
with, especially considering the demonstrative value of robust code.

Signed-off-by: Kerin Millar <kfm@plushkava.net>
Bug: https://bugs.gentoo.org/881383
Signed-off-by: Sam James <sam@gentoo.org>