media-print/brlaser: move to media-print and add new version
The old brlaser is not supported since years. As I'm gonna take
one of these printers I've took a look over it and updated a bit
adding new printers and some code-checks (thanks GH actions).
Let me know what you think about that update and if it is good for
you.

Signed-off-by: Marco Scardovi <scardracs-gentoo@proton.me>