x11-drivers/nvidia-drivers: bump NV_KERNEL_MAX to 6.1
All branches build fine with it, tested runtime a bit with 525
and "seems" fine at least for it too (not that I really use new
kernel branches until .5+ and gentoo-kernel bump).

For the now-unsupported 390 and if unpatched, that means planned
LTS 6.1 will likely be the kernel to stay on (forever) when 5.15
reaches end of life.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>